requirements.txt
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.2
django-filter==2.0.0
django-rest-swagger==2.1.1
djangorestframework==3.8.2
idna==2.7
itypes==1.1.0
Jinja2==2.10
jsondiff==1.1.2
jsonfield==2.0.2
jsonschema==2.6.0
Markdown==3.0.1
MarkupSafe==1.0
openapi-codec==1.3.2
psycopg2-binary==2.7.5
pytz==2018.5
requests==2.19.1
simplejson==3.16.0
uritemplate==3.0.0
urllib3==1.23
